chore: remove outdated benchmark files and consolidate benchmark infrastructure into a single file for improved organization and maintainability
